🔥🐎 Move obsolete code to obsolete files and update perf and tests
