+++ src/Yarhl.IntegrationTests/PluginDiscovery.cs
@@ -23,6 +23,7 @@ namespace Yarhl.IntegrationTests
     using System.IO;
     using System.Linq;
     using NUnit.Framework;
+    using Yarhl.Plugins;
 
     [TestFixture]
     public class PluginDiscovery
