+++ src/Yarhl.IntegrationTests/Yarhl.IntegrationTests.csproj
@@ -7,12 +7,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Yarhl\Yarhl.csproj" />
-
     <PackageReference Include="NUnit" />
     <PackageReference Include="NUnit3TestAdapter" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="coverlet.collector" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Yarhl.Plugins\Yarhl.Plugins.csproj" />
   </ItemGroup>
 
 </Project>
