+++ src/Yarhl.UnitTests/Plugins/ConverterFindableTests.cs
@@ -17,12 +17,15 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Yarhl.UnitTests.FileFormat
+
+namespace Yarhl.UnitTests.Plugins
 {
     using System;
     using System.Linq;
     using NUnit.Framework;
     using Yarhl.FileFormat;
+    using Yarhl.Plugins;
+    using Yarhl.UnitTests.FileFormat;
 
     [TestFixture]
     public class ConverterFindableTests
