+++ src/Yarhl.UnitTests/Plugins/ConverterMetadataTests.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 SceneGate
+﻿// Copyright (c) 2019 SceneGate
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -17,12 +17,12 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Yarhl.UnitTests.FileFormat
+namespace Yarhl.UnitTests.Plugins
 {
     using System;
     using System.Diagnostics.CodeAnalysis;
     using NUnit.Framework;
-    using Yarhl.FileFormat;
+    using Yarhl.Plugins.FileFormat;
 
     [TestFixture]
     public class ConverterMetadataTests
