+++ src/Yarhl.UnitTests/Yarhl.UnitTests.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Yarhl.Plugins\Yarhl.Plugins.csproj" />
     <ProjectReference Include="..\Yarhl\Yarhl.csproj" />
     <ProjectReference Include="..\Yarhl.Media.Text\Yarhl.Media.Text.csproj" />
