+++ src/Yarhl/FileFormat/ConverterMetadata.Obsolete4.cs
