+++ src/Yarhl/FileFormat/FormatMetadata.Obsolete4.cs
