+++ src/Yarhl/FileFormat/IExportMetadata.Obsolete4.cs
