+++ src/Yarhl/FileFormat/IInitializer.Obsolete4.cs
