+++ src/Yarhl/IO/SeekMode.Obsolete4.cs
