+++ src/Yarhl/PluginManager.Obsolete4.cs
